+++ CHANGES.rst
@@ -15,6 +15,7 @@ Major Changes
 - build(pkg): apply dependency upgrades & new ruff rules
 - build(pkg): test & signal support for python 3.14
 - feat(slots): provide a `clear` method for containers
+- feat(slots): pass (un)mapped attributes down class hierarchy
 - fix(db): ensure manifest's metatag field is populated